+++ kernel/arch/aarch64/trap/trap.cpp
@@ -13,6 +13,7 @@
 #include "timer/timer.h"
 #include "sched/sched.h"
 #include "sched/task.h"
+#include "mm/mm.h"
 
 // Forward declaration of syscall dispatch
 extern "C" void stlx_aarch64_syscall_dispatch(aarch64::trap_frame* tf);
@@ -77,6 +78,19 @@ void stlx_aarch64_el0_sync_handler(aarch64::trap_frame* tf) {
         stlx_aarch64_syscall_dispatch(tf);
         restore_post_trap_elevation_state();
         return;
+    }
+
+    // If it's a page fault, attempt to handle it for on-demand paging
+    if (in_user_code && (
+        ec == aarch64::EC_DATA_ABORT_LOWER ||
+        ec == aarch64::EC_INST_ABORT_LOWER)
+    ) {
+        uintptr_t fault_addr = aarch64::get_far(tf);
+        if (mm::handle_user_pf(guard.task_core->mm_ctx, fault_addr, esr)) {
+            // Fault has been handled successfully, restart instruction
+            restore_post_trap_elevation_state();
+            return;
+        }
     }
 
     if (in_user_code && (
